Restructure folder structure to work with AOT
