--- src/client/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "angularCompilerOptions": {
-    "entryModule": "./client-app.module#ClientAppModule"
-  }
-}
