+++ src/main.client.ts
@@ -3,7 +3,7 @@ import 'reflect-metadata';
 import 'rxjs/Observable';
 import 'rxjs/add/operator/map';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
-import { ClientAppModule } from './client-app.module';
+import { ClientAppModule } from './app/client-app.module';
 import { bootloader } from '@angularclass/hmr';
 
 export function main() {
