+++ src/tsconfig.client.json
@@ -1,0 +1,6 @@
+{
+  "extends": "../tsconfig.json",
+  "angularCompilerOptions": {
+    "entryModule": "./app/client-app.module#ClientAppModule"
+  }
+}
