+++ src/tsconfig.server.json
@@ -1,0 +1,6 @@
+{
+  "extends": "../tsconfig.json",
+  "angularCompilerOptions": {
+    "entryModule": "./app/server-app.module#ServerAppModule"
+  }
+}
